Speed up PAN12 testing method
The most time consuming part of the algorithm testing is the tagging of
the samples. Herewith we improve the speed when testing the same
problem considering different numbers of the most common words. The
former implementation was quite inefficient. Now, we need to tag only
once per file and problem.